% Generated by roxygen2 (4.0.2): do not edit by hand
\name{A.design.plot}
\alias{A.design.plot}
\title{Plot number of reads in each sample for each grouping factor.}
\usage{
A.design.plot(fqc, design.table)
}
\arguments{
\item{fqc}{FastQA from package ShortRead}

\item{design.table}{data.frame holds information about experimantal design}
}
\value{
list of plot objects
}
\description{
Bar plots representing number of reads in each sample.
There is one bar plot per grouping factor.
Samples are grouped and clored by grouping factor.
Grouping factors are extracted from design table.
Samples are expected to be separate fastq files.
}

