% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Aplot}
\alias{Aplot}
\title{Plot number of reads in each sample.}
\usage{
Aplot(fqc)
}
\arguments{
\item{fqc}{FastQA from package ShortRead}
}
\value{
plot object
}
\description{
Bar plot representing number of reads in each sample.
Samples are expected to be separate fastq files.
}

