\name{Aplot2}
\alias{Aplot2}
\title{Plot number of reads in each sample.}
\usage{
  Aplot2(fqc)
}
\arguments{
  \item{fqc}{FastQA from package ShortRead}
}
\value{
  plot object
}
\description{
  Bar plot representing number of reads in each sample.
  Samples are expected to be separate fastq files.
}

