% Generated by roxygen2 (4.0.0): do not edit by hand
\name{B1.design.plot}
\alias{B1.design.plot}
\title{Plot read lenght distribution}
\usage{
B1.design.plot(samples, design.table)
}
\arguments{
\item{samples}{ShortReadQ object from package ShortRead}

\item{design.table}{data.frame holds information about experimantal design}
}
\value{
list of plot objects
}
\description{
Density plot representing freaquencies or readlenths.
There is one plot per grouping factor.
Samples clored by grouping factor.
Grouping factors are extracted from design.table.
Samples are explected to be separate fastq files.
}

