% Generated by roxygen2 (4.0.0): do not edit by hand
\name{B1plot}
\alias{B1plot}
\title{Plot read lenght distribution}
\usage{
B1plot(samples)
}
\arguments{
\item{samples}{ShortReadQ object from package ShortRead}
}
\value{
plot object
}
\description{
Density plot representing freaquencies or readlenths.
Samples are explected to be separate fastq files.
}

