\name{B1plot2}
\alias{B1plot2}
\title{Plot read lenght distribution}
\usage{
  B1plot2(samples)
}
\arguments{
  \item{fqc}{FastQA from package ShortRead}
}
\value{
  plot object
}
\description{
  Density plot representing freaquencies or readlenths.
  Samples are explected to be separate fastq files.
}

