% Generated by roxygen2 (4.0.0): do not edit by hand
\name{B2.design.plot}
\alias{B2.design.plot}
\title{Plot fraction of reads with particular lengh or longer.}
\usage{
B2.design.plot(samples, design.table)
}
\arguments{
\item{samples}{ShortReadQ object from package ShortRead}

\item{design.table}{data.frame holds information about experimantal design}
}
\value{
plots list of plot objects
}
\description{
Line plot showing fraction of reads on y axis
and minimal read length on x axis.
One line per sample.
There is one plot per grouping factor.
Samples clored by grouping factor.
Grouping factors are extracted from design.table.
Samples are explected to be separage fastq files.
}

