\name{B2plot2}
\alias{B2plot2}
\title{Plot fraction of reads with particular lengh or longer.}
\usage{
  B2plot2(samples)
}
\arguments{
  \item{samples}{ShortReadQ object from package ShortRead}
}
\value{
  plot object
}
\description{
  Line plot showing fraction of reads on y axis and minimal
  read length on x axis. One line per sample. Samples are
  explected to be separage fastq files.
}

