% Generated by roxygen2 (4.0.0): do not edit by hand
\name{C1.design.plot}
\alias{C1.design.plot}
\title{Plot mean read quality distribution per sample.}
\usage{
C1.design.plot(samples, design.table)
}
\arguments{
\item{samples}{ShortReadQ object from package ShortRead}

\item{design.table}{data.frame holds information about experimantal design}
}
\value{
list of plot objects
}
\description{
Boxplot showing the distribution of mean read quality.
One boxplot per sample.
There is one plot per grouping factor.
Samples clored by grouping factor.
Grouping factors are extracted from design.table.
Samples are explected to be separage fastq files.
}

