\name{C1plot}
\alias{C1plot}
\title{Plot mean read quality distribution per sample.}
\usage{
C1plot(samples)
}
\arguments{
  \item{samples}{ShortReadQ object from package ShortRead}
}
\value{
plot object
}
\description{
Boxplot showing the distribution of mean read quality. One
boxplot per sample. Samples are explected to be separage
fastq files.
}

