\name{C1plot2}
\alias{C1plot2}
\title{Plot mean read quality distribution per sample.}
\usage{
  C1plot2(samples)
}
\arguments{
  \item{samples}{ShortReadQ object from package ShortRead}
}
\value{
  plot object
}
\description{
  Boxplot showing the distribution of mean read quality.
  One boxplot per sample. Samples are explected to be
  separage fastq files.
}

