% Generated by roxygen2 (4.0.2): do not edit by hand
\name{C2.design.plot}
\alias{C2.design.plot}
\title{Plot mean base quality at particualar position in the read.}
\usage{
C2.design.plot(samples, design.table)
}
\arguments{
\item{samples}{ShortReadQ object from package ShortRead}

\item{design.table}{data.frame holds information about experimantal design}
}
\value{
list of plot objects
}
\description{
Line plot showing mean quality of the bases per position in the read.
One line per sample.
Samples are explected to be separage fastq files.
There is one plot per grouping factor.
Samples colored by grouping factor.
Grouping factors are extracted from design.table.
}

