% Generated by roxygen2 (4.0.2): do not edit by hand
\name{C2plot}
\alias{C2plot}
\title{Plot mean base quality at particualar position in the read.}
\usage{
C2plot(samples)
}
\arguments{
\item{samples}{ShortReadQ object from package ShortRead}
}
\value{
plot object
}
\description{
Line plot showing mean quality of the bases per position in the read.
One line per sample.
Samples are explected to be separage fastq files.
}

