% Generated by roxygen2 (4.0.0): do not edit by hand
\name{C3plot}
\alias{C3plot}
\title{Plot fraction of reads with partucular quality or higher per position.}
\usage{
C3plot(fqc, samples)
}
\arguments{
\item{fqc}{FastQA from package ShortRead}

\item{samples}{ShortReadQ object from package ShortRead}
}
\value{
plot object
}
\description{
4 line plots, for qualities 18, 20, 24 and 28, showing fraction
of reads with one of those qualities per position in the read.
One line per sample.
Samples are explected to be separage fastq files.
}

