% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CountTable}
\alias{CountTable}
\title{Count reads in all}
\usage{
CountTable(bam_dir, base_name, table_path)
}
\arguments{
\item{bam_dir}{directory with alignments}

\item{base_name}{part of the name commot to all the alignment files}

\item{table_path}{file path for resulting count table}
}
\description{
Count reads in all
}

