% Generated by roxygen2 (4.0.2): do not edit by hand
\name{D1plot}
\alias{D1plot}
\title{Plot base frequnecy for first 30 nt.}
\usage{
D1plot(samples)
}
\arguments{
\item{samples}{ShortReadQ object from package ShortRead}
}
\value{
plot object
}
\description{
Linepolots representing frequencies of bases, one line per base,
for first 30 nt in the read (from 5')
One plot per sample.
Each sample is a fastq files.
}

