% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalise counts based on normalisation spike counts}
\usage{
normalize_counts(spike_counts, counts, norm_counts)
}
\arguments{
\item{spike_counts}{file with count table of normalization spikes}

\item{counts}{file with counts table to normalize}

\item{norm_counts}{file that will be used for calculated normalized counts}
}
\description{
Normalise counts based on normalisation spike counts
}

