\name{plot_mm_dis}
\alias{plot_mm_dis}
\title{Plot distribution of mismatches and indels in reads.}
\usage{
  plot_mm_dis(bamfile, plotfile)
}
\arguments{
  \item{bamfile}{string bamfile filepath}

  \item{plotfile}{string output png filepath}
}
\description{
  Plot distribution of mismatches and indels in reads.
}

