\name{PlotMeanReadQuality}
\alias{PlotMeanReadQuality}
\title{Plot mean quality per read}
\usage{
  PlotMeanReadQuality(data.dir, file.pattern)
}
\arguments{
  \item{data.dir}{path to data dir}

  \item{file.pattern}{pattern of input fastq}
}
\value{
  plot object
}
\description{
  Plot mean quality per read
}

