\name{PlotMMDist}
\alias{PlotMMDist}
\title{Plot distribution of mismatches and indels in reads.}
\usage{
  PlotMMDist(bamfile, plotfile)
}
\arguments{
  \item{bamfile}{string bamfile filepath}

  \item{plotfile}{string output png filepath}
}
\description{
  Plot distribution of mismatches and indels in reads.
}

