% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotNormalSpikeCounts}
\alias{PlotNormalSpikeCounts}
\title{Plot normalised spike counts}
\usage{
PlotNormalSpikeCounts(counts_file, total_reads)
}
\arguments{
\item{counts_file}{csv file with spike counts}

\item{sample_read_couts}{named list, number of reads per each sample}
}
\description{
Plot normalised spike counts
}

