\name{PlotPerCycleBaseCalls}
\alias{PlotPerCycleBaseCalls}
\title{Plot base call per cycle}
\usage{
PlotPerCycleBaseCalls(data.dir, file.pattern, type = "fastq")
}
\arguments{
  \item{data.dir}{string path to data}

  \item{file.pattern}{string pattern for input files}

  \item{type}{stirng input type, 'fastq' is default value}
}
\value{
plot
}
\description{
Plot base call per cycle
}

