\name{PlotPerCycleCumQuality}
\alias{PlotPerCycleCumQuality}
\title{Plot cumulative qualities}
\usage{
  PlotPerCycleCumQuality(data.dir, file.pattern)
}
\arguments{
  \item{data.dir}{string path to data}

  \item{file.pattern}{string pattern for input files}
}
\value{
  plot
}
\description{
  Plot cumulative qualities
}

