\name{PlotPerCycleQuality}
\alias{PlotPerCycleQuality}
\title{Plot quality scores per cycle}
\usage{
  PlotPerCycleQuality(data.dir, file.pattern,
    type = "fastq")
}
\arguments{
  \item{data.dir}{string path to data}

  \item{file.pattern}{string pattern for input files}

  \item{type}{stirng input type, 'fastq' is default value}
}
\value{
  plot
}
\description{
  Plot quality scores per cycle
}

