% Generated by roxygen2 (4.0.0): do not edit by hand
\name{PlotReadLengthDistribution}
\alias{PlotReadLengthDistribution}
\title{Plot read length distribution}
\usage{
PlotReadLengthDistribution(data.dir, file.pattern)
}
\arguments{
\item{data.dir}{path to data dir}

\item{file.pattern}{pattern of input fastq}
}
\value{
plot object
}
\description{
Plot read length distribution
}

