\name{Aplot}
\alias{Aplot}
\title{Plot number of reads in each sample in an experiment}
\usage{
  Aplot(data.dir, file.pattern)
}
\arguments{
  \item{data.dir}{string path to data}

  \item{file.pattern}{string pattern for input files}
}
\value{
  plot object
}
\description{
  Plot number of reads in each sample in an experiment
}

