% Generated by roxygen2 (4.0.0): do not edit by hand
\name{PlotAlignAnnot}
\alias{PlotAlignAnnot}
\title{Plot aligned reads over annotated region.}
\usage{
PlotAlignAnnot(aln, annots, plotfile)
}
\arguments{
\item{aln}{read alignments in GAlignments objet (GenomicTRanges)}

\item{annots}{genome annotations in GRanges object}

\item{plotfile}{path for output plot}
}
\description{
Plot aligned reads over annotated region.
}

