\name{PlotCoverageAnnot}
\alias{PlotCoverageAnnot}
\title{Plot aligned reads over annotated region.}
\usage{
  PlotCoverageAnnot(aln, annots, plotfile)
}
\arguments{
  \item{aln}{read alignments in GAlignments objet
  (GenomicTRanges)}

  \item{annot}{genome annotations in GRanges object}

  \item{plotfile}{path for output plot}
}
\description{
  Plot aligned reads over annotated region.
}

