% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotMMDist}
\alias{PlotMMDist}
\title{Plot distribution of mismatches and indels in alignment.}
\usage{
PlotMMDist(bamfile, plotfile)
}
\arguments{
\item{bamfile}{string bamfile filepath}

\item{plotfile}{string output png filepath}
}
\description{
Plot distribution of mismatches and indels in alignment.
}

