\name{PlotPerCycleQuality}
\alias{PlotPerCycleQuality}
\title{Plot quality scores per cycle}
\usage{
  PlotPerCycleQuality(data.dir, file.pattern)
}
\arguments{
  \item{data.dir}{string path to data}

  \item{file.pattern}{string pattern for input files}
}
\value{
  plot
}
\description{
  Plot quality scores per cycle
}

