% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotPerCycleQuality}
\alias{PlotPerCycleQuality}
\title{Plot quality scores per cycle}
\usage{
PlotPerCycleQuality(data.dir, file.pattern)
}
\arguments{
\item{data.dir}{string path to data}

\item{file.pattern}{string pattern for input files}
}
\value{
plot object
}
\description{
Plot quality scores per cycle
}

